Remote unused deps, fix hot reloading
