+++ .babelrc
@@ -1,8 +1,7 @@
 {
   "presets": ["es2015", "react"],
   "plugins": [
-    "transform-object-rest-spread",
-    "react-hot-loader/babel"
+    "transform-object-rest-spread"
   ],
   "env": {
     "production": {
@@ -12,27 +11,8 @@
       ]
     },
     "development": {
-      "presets": ["react-hmre"],
       "plugins": [
-        [
-          "react-transform",
-            {
-            "transforms": [
-              {
-                "transform": "react-transform-hmr",
-                "imports": ["react"],
-                "locals": ["module"]
-              },
-              {
-                "transform": "react-transform-render-visualizer"
-              },
-              {
-                "transform": "react-transform-catch-errors",
-                "imports": ["react","redbox-react"]
-              }
-            ]
-          }
-        ]
+        "react-hot-loader/babel"
       ]
     }
   }
