+++ package.json
@@ -118,9 +118,6 @@
     "phantomjs-prebuilt": "^2.1.14",
     "react-addons-test-utils": "^15.0.0",
     "react-hot-loader": "3.0.0-beta.6",
-    "react-transform-catch-errors": "^1.0.2",
-    "react-transform-hmr": "^1.0.2",
-    "react-transform-render-visualizer": "^0.4.0",
     "redbox-react": "^1.2.2",
     "redux-devtools": "^3.2.0",
     "redux-devtools-chart-monitor": "^1.2.0",
