+++ webpack/webpack.config.js
@@ -1,5 +1,7 @@
 /* globals __dirname */
 const path = require('path');
+const fs = require('fs');
+const merge = require('lodash/merge');
 
 const baseContext = path.join(__dirname, '../client');
 const environmentConext = path.join(__dirname, '../environment');
@@ -7,7 +9,7 @@ const universalContext = path.join(__dirname, '../universal');
 
 module.exports = {
   context: baseContext,
-  entry: '../client/application.jsx',
+  entry: '../client/index.jsx',
   output: {
     path: path.resolve(__dirname, '../server', 'public', 'dist'),
     filename: 'bundle.js',
@@ -25,7 +27,14 @@ module.exports = {
       {
         test: /\.js(x|)?$/,
         use: {
-          loader: 'babel-loader?plugins[]=transform-object-rest-spread'
+          loader: 'babel-loader',
+
+          // Work-around to enable commonjs on server and native modules on client,
+          // while (mostly) maintaining one source for Babel configuration
+          options: merge(JSON.parse(fs.readFileSync('./.babelrc')), {
+            babelrc: false,
+            presets: [['es2015', { modules: false }], 'react']
+          })
         },
         include: [baseContext, universalContext],
         exclude: /node_modules/
